+++ JAVA/controller/src/main/java/interfaces/ICategories.java
@@ -7,7 +7,8 @@ import java.util.ArrayList;
 public interface ICategories {
     ArrayList<String> getCategories();
     ArrayList<Long> getCategoriesId();
-    CategoryBO getCategory(int id);
+    CategoryBO getCategory(long id);
+
     void addCategory(String name);
     boolean updateCategory(String name, String newName);
     boolean deleteCategory(String name);
